Implemented locking of directories in the main storage.
git-svn-id: http://manti.exp.bessy.de/svn/icat/ids-plugin-hzb/trunk@708 8efdbd46-c5fb-49ab-9956-99f62928ec21

+++ src/main/java/de/helmholtz_berlin/icat/ids/storage/TreeDeleteVisitor.java
@@ -12,7 +12,12 @@ public class TreeDeleteVisitor extends SimpleFileVisitor<Path> {
     @Override
     public FileVisitResult visitFile(Path file, BasicFileAttributes attrs) 
 	throws IOException {
-	Files.delete(file);
+	// Make sure that a lock file, if present, is the very last
+	// thing that gets deleted in a directory.  Don't delete it
+	// here, but rather in postVisitDirectory().
+	if (! file.getFileName().toString().equals(".lock")) {
+	    Files.delete(file);
+	}
 	return FileVisitResult.CONTINUE;
     }
 
@@ -20,6 +25,11 @@ public class TreeDeleteVisitor extends SimpleFileVisitor<Path> {
     public FileVisitResult postVisitDirectory(Path dir, IOException e) 
 	throws IOException {
 	if (e == null) {
+	    // Delete a lock file if present, see above.
+	    Path lockf = dir.resolve(".lock");
+	    if (Files.exists(lockf)) {
+		Files.delete(lockf);
+	    }
 	    Files.delete(dir);
 	    return FileVisitResult.CONTINUE;
 	} else {
